nuke-references: get rid of perl
Since bash does not have negative lookaheads in it's regex engine
(or lookaheads at all really), the bash implementation loads one line
into memory at a time and scans the same line for store references
until all has been dealt with by nuking or ignoring.

To replace the `\Q` and `\E` escaping mechanism, single tick quotations are
used in the regexes so bash considers the contents of `@storePath@` to be literal,
and the provided `@storePath@` have escaped any single ticks in `builtins.storePath`
(but please don't use `'` in the path to the nix store for real...).

This variant is likely slower and more possibly more memory hungry
(especially for binaries where it could be far between newlines)
than the implementation using perl, and any performance benefits would
have to come from added build parallelism with the removed dependency constraint.